requirements.txt
pytest==8.3.4
allure-pytest~=2.13.5
requests~=2.32.3
selenium~=4.27.1
webdriver-manager~=4.0.2